requirements.txt
flask==2.3.3
markdown==3.5
python-dotenv==1.0.0
langchain==0.3.17
langchain-openai==0.3.3
langchain-community==0.3.16